# gin/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/selectByArtType": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get UniversityByArtType",
                "parameters": [
                    {
                        "type": "string",
                        "description": "School_name",
                        "name": "school_name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/selectByCooperateType": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get UniversityByCooperateType",
                "parameters": [
                    {
                        "type": "string",
                        "description": "School_name",
                        "name": "school_name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/selectByFinanceType": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get UniversityByFinanceType",
                "parameters": [
                    {
                        "type": "string",
                        "description": "School_name",
                        "name": "school_name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/selectByLanguageType": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get UniversityByLanguageType",
                "parameters": [
                    {
                        "type": "string",
                        "description": "School_name",
                        "name": "school_name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/selectByLawType": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get UniversityByLawType",
                "parameters": [
                    {
                        "type": "string",
                        "description": "School_name",
                        "name": "school_name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/selectByMedicineType": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get UniversityByMedicineType",
                "parameters": [
                    {
                        "type": "string",
                        "description": "School_name",
                        "name": "school_name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/selectByMinBanType": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get UniversityByMinBanType",
                "parameters": [
                    {
                        "type": "string",
                        "description": "School_name",
                        "name": "school_name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/selectByMinZuType": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get UniversityByMinZuType",
                "parameters": [
                    {
                        "type": "string",
                        "description": "School_name",
                        "name": "school_name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/selectBySportsType": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get UniversityBySportsType",
                "parameters": [
                    {
                        "type": "string",
                        "description": "School_name",
                        "name": "school_name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/selectByZongBang": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get UniversityByZongBang",
                "parameters": [
                    {
                        "type": "string",
                        "description": "School_name",
                        "name": "school_name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/selectUniversity": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get University",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "app.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}